Third pass
Get rid of the calls to `MessageBuilder::new()->setcontent()` in favor of letting logEvent handle it. This will make it easier in the future to allow for overriding the function to change the output, such as sending the content as an embed instead of raw message content. Passing the name of the event will also make it easier for more precise conditional logic.